% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/london_display2.R
\name{london_display}
\alias{london_display}
\title{Display available palettes}
\usage{
london_display(name = NULL, n, colourblind = FALSE, direction = 1)
}
\arguments{
\item{name}{\code{character}. Name of the palette. Either:
\code{Cholmondeley}, \code{Copley}, \code{Blake}, \code{Turner},
\code{Millais}, \code{Sargent}, \code{Waterhouse}, \code{Grant},
\code{Lowry}, \code{Bacon}, \code{Hockney}. If NULL (default),
all available palettes are displayed.}

\item{n}{\code{integer}. Number of requested colours. If no value is
provided, the default number of colours are returned.}

\item{colourblind}{\code{logical}. Should only colourblind friendly palettes
be returned?}

\item{direction}{\code{integer}. Sets the orders of colours in the scale.
Default order is 1. Palette can be reversed by specifying -1.}
}
\value{
Palettes
}
\description{
A function for displaying the available palettes in \code{london}.
}
\details{
Palettes
}
\examples{
london_display(name = "Hockney")
}
