% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/london_palette.R
\name{london_palette}
\alias{london_palette}
\title{Generate colour palette}
\usage{
london_palette(name, n, type = "discrete", direction = 1, display = FALSE)
}
\arguments{
\item{name}{\code{character}. Name of desired palette. Either:
\code{Cholmondeley}, \code{Copley}, \code{Blake}, \code{Turner},
\code{Millais}, \code{Sargent}, \code{Waterhouse}, \code{Grant},
\code{Lowry}, \code{Bacon}, \code{Hockney}.}

\item{n}{\code{numeric}. Number of requested colours. If no value is
provided, the default number of colours is returned.}

\item{type}{\code{character}. Type of colour scale.
Either "continuous" or "discrete". Use "continuous" if you want
to automatically interpolate between colours.}

\item{direction}{\code{numeric}. Sets the orders of colours in the scale.
Default order is 1. Palette is reversed by specifying -1.}

\item{display}{\code{logical}. Should the colour palette be displayed?}
}
\value{
A vector of the hex colour codes of the palette.
}
\description{
A function to generate a colour palette based on the available palettes in
\code{london}.
}
\examples{
london_palette(name = "Hockney")
}
\keyword{colours}
