% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/london_sample.R
\name{london_sample}
\alias{london_sample}
\title{Generate colour palette from image}
\usage{
london_sample(path, n = 4, type = "discrete", direction = 1, plot = TRUE)
}
\arguments{
\item{path}{The file path or url of the image you wish to sample colours from.}

\item{n}{The number of colours to sample from the image. If n is not provided, the length of the palette defaults to 4.}

\item{type}{Specify either "discrete" or "continuous". If "continuous", three colours are sampled from the image and automatically interpolated between. Defaults to "discrete".}

\item{direction}{Specify the order of colours. Defaults to 1. If direction is -1, palette is reversed.}

\item{plot}{(Logical) Plot sampled colour palette. Defaults to TRUE.}
}
\description{
This function allows you to generate a colour palette from an image.
}
\examples{
url <- "https://t4.ftcdn.net/jpg/02/90/91/11/240_F_290911157_rJMt9bp5sq64hyY6TklM9XS5GoDg07Vh.jpg"
london_sample(path = url, n = 4, type = "discrete")

}
