% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_colour_tate_d.R
\name{scale_colour_tate_d}
\alias{scale_colour_tate_d}
\title{Discrete tate palettes for plotting with ggplot2}
\usage{
scale_colour_tate_d(name, direction = 1, ...)
}
\arguments{
\item{name}{\code{character}. Name of desired palette. Either:
\code{Cholmondeley}, \code{Copley}, \code{Blake}, \code{Turner},
\code{Millais}, \code{Sargent}, \code{Waterhouse}, \code{Grant},
\code{Lowry}, \code{Bacon} or \code{Hockney}.}

\item{direction}{\code{numeric}. Sets the order of colours in the scale.
Default order is 1. Palette is reversed by specifying -1.}

\item{...}{Other arguments passed on to
\code{\link[ggplot2]{scale_colour_gradientn}}}
}
\description{
A function to use colour palettes from \code{tate} with \code{ggplot2}. Use
\code{\link{scale_colour_met_d}} and \code{\link{scale_fill_met_d}} for
discrete scales, and \code{\link{scale_colour_met_c}} and
\code{\link{scale_fill_met_c}} for continuous scales.
}
\examples{
# Load libraries
library(ggplot2)
library(palmerpenguins)
# Plot data
ggplot(data = penguins,
       aes(x = bill_length_mm, y = bill_depth_mm, colour = species)) +
  geom_point() +
  scale_colour_tate_d("Millais", direction = 1)
}
