% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tate_construct.R
\name{tate_construct}
\alias{tate_construct}
\title{Construct a tate palette object.}
\usage{
tate_construct(x, name)
}
\arguments{
\item{x}{\code{vector/list}. A one-component vector or list with each
element containing the palette colours.}

\item{name}{\code{character}. Name(s) of the palette(s).}
}
\value{
A \code{tate_palette} or \code{tate_palette_list} object.
}
\description{
Constructs a \code{tate_palette} or \code{tate_palette_list} object.
}
