% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tate_display.R
\name{tate_display}
\alias{tate_display}
\title{Display palette}
\usage{
tate_display(x)
}
\arguments{
\item{x}{\code{tate_palette} or \code{tate_palette_list}.
These objects are output by the
\code{\link[tate:tate_palette]{tate::tate_palette()}} or
\code{\link[tate:tate_sample]{tate::tate_sample()}} functions or can
be constructed via the
\code{\link[tate:tate_construct]{tate::tate_construct()}}.}
}
\value{
A plot of the input palette(s).
}
\description{
A function for visualising palettes.
}
\examples{
Hockney <- tate_palette(name = "Hockney")
tate_display(x = Hockney)
all_palettes <- tate_palette(name = "All")
tate_display(x = all_palettes)
}
