% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tate_palette.R
\name{tate_palette}
\alias{tate_palette}
\title{Generate colour palette}
\usage{
tate_palette(
  name,
  n = NULL,
  type = "discrete",
  direction = 1,
  friendly = FALSE,
  display = FALSE
)
}
\arguments{
\item{name}{\code{character}. Name of desired palette. Either:
\code{Cholmondeley}, \code{Copley}, \code{Blake}, \code{Turner},
\code{Millais}, \code{Sargent}, \code{Waterhouse}, \code{Grant},
\code{Lowry}, \code{Bacon}, \code{Hockney} or \code{All}, if all are desired.}

\item{n}{\code{numeric}. Number of requested colours. If no value is
provided, the default number of colours is returned.}

\item{type}{\code{character}. Type of colour scale.
Either "continuous" or "discrete". Use "continuous" if you want
to automatically interpolate between colours.}

\item{direction}{\code{numeric}. Sets the orders of colours in the scale.
Default order is 1. Palette is reversed by specifying -1.}

\item{friendly}{\code{logical}. Should non-colourblind friendly palettes
be removed?}

\item{display}{\code{logical}. Should the colour palette be displayed?}
}
\value{
A vector of hex colour codes.
}
\description{
A function to generate a colour palette based on the available palettes in
\code{tate}.
}
\examples{
tate_palette(name = "Hockney", display = TRUE)
tate_palette(name = "Sargent", n = 1000, type = "continuous", display = TRUE)
tate_palette(name = "All", display = TRUE)
tate_palette(name = "All", friendly = TRUE, display = TRUE)
}
\keyword{colours}
